requirements.txt
streamlit==1.32.0
opencv-python==4.8.1.78
numpy==1.26.2
pillow==10.1.0
tensorflow>=2.5.0